lmb: fix "implicit declaration of function 'lmb_free'" warning

+++ lib_ppc/bootm.c
@@ -212,7 +212,7 @@ do_bootm_linux(cmd_tbl_t *cmdtp, int flag, int argc, char *argv[],
 		}
 
 		/* Delete the old LMB reservation */
-		lmb_free(lmb, of_flat_tree, fdt_totalsize(of_flat_tree));
+		lmb_free(lmb, (uint64_t)of_flat_tree, fdt_totalsize(of_flat_tree));
 
 		/* Calculate the actual size of the fdt */
 		actualsize = fdt_off_dt_strings(of_flat_tree) +
